+++ src/main/java/UI/Finances.java
@@ -1,4 +1,4 @@
-package UI_Controllers;
+package UI;
 
 public class Finances {
 }
