+++ src/main/java/UI/HomeController.java
@@ -1,0 +1,10 @@
+package UI;
+
+import java.io.IOException;
+
+public class HomeController {
+
+    public void ButtonClicked(javafx.event.ActionEvent actionEvent) throws IOException {
+        UI.StartUI.IncomeLoad();
+    }
+}
