+++ src/main/resources/UI/Expense.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="UI_Controllers.Finances"
+            fx:controller="UI.Expense"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
