+++ src/main/resources/UI/Income.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.text.Text?>
 
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="UI_Controllers.IncomeController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="UI.IncomeController">
    <children>
       <Text layoutX="239.0" layoutY="196.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Income menu" />
    </children>
