+++ src/main/resources/UI/home-ui.fxml
@@ -4,9 +4,9 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
-<?xml-stylesheet type="text/css" href="home.css"?>
+<!--<?xml-stylesheet type="text/css" href="home.css"?>-->
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="UI_Controllers.HomeController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="UI.HomeController">
    <children>
       <Button id="Income_Btn" layoutX="219.0" layoutY="140.0" mnemonicParsing="false" prefHeight="50.0" prefWidth="162.0" text="Income" onAction="#ButtonClicked">
          <font>
